refactor: change DualQuaternion folder to dual_quaternion
