+++ source/state_representation/include/state_representation/space/dual_quaternion/DualQuaternionPose.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "state_representation/space/DualQuaternion/DualQuaternionState.hpp"
+#include "state_representation/space/dual_quaternion/DualQuaternionState.hpp"
 
 namespace state_representation {
 /**
