+++ source/state_representation/include/state_representation/space/dual_quaternion/DualQuaternionState.hpp
