+++ source/state_representation/tests/testDualQuaternionState.cpp
@@ -1,5 +1,4 @@
-#include "state_representation/space/DualQuaternion/DualQuaternionPose.hpp"
-#include "state_representation/space/DualQuaternion/DualQuaternionTwist.hpp"
+#include "state_representation/space/dual_quaternion/DualQuaternionPose.hpp"
 #include <gtest/gtest.h>
 
 TEST(MultiplyTransformsBothOperators, PositiveNos) {
